fixed task 2 compilation warning

+++ 2-selection_sort.c
@@ -12,7 +12,7 @@ void selection_sort(int *array, size_t size)
 
 	for (i = 0; i < size; i++)
 	{
-		int idxOfMin = i, j;
+		size_t idxOfMin = i, j;
 
 		for (j = i + 1; j < size; j++) /* find min num in remaining numbers */
 		{
